pyproject.toml
[tool.poetry]
name = "django-library-project-test"
version = "0.1.0"
description = ""
authors = ["Julia-Gareeva <uliastekolnikova12822@gmail.com>"]
readme = "README.md"
packages = [{include = "django_library_project_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
